Restart .vimrc
Main intent of this was to make vim startup time faster, so I've cleared
a lot of unused plugins, cut down used but not very important ones (like
Airline). Also I've switched to vim-plug because it's capable of loading
plugins on-demand.

I've switched back to Neovim because of Neomake and Deoplete, but still
trying to maintain compatibility with vim (by using `if has('nvim')`
checks)