Get rid of vim-gitgutter
While it's nice to see which lines are changed, it interferes with my
workflow for a bit: I rely on Neomake to make sure that file has no
syntax errors, and when file doesn't fit the screen the quick way I can
tell that there is no more errors is by looking at SignColumn - if it's
hidden then there is no errors. But vim-gitgutter also uses SignColumn
to show it's signs, ruining this workflow for me.