Add RuboCop configuration
My goal in tweaking rules is that I don't develop a habit of ignoring
RuboCop's warnings, so I'm disabling those which are not important to me.

- `FrozenStringLiteralComment` is disabled, because Ruby 3 isn't going to
be released soon and it's really tedious to add this comment to every
file.

- `Documentation` requires each class or module to have top-level
documentation, which I often don't do, especially when working on Ruby
on Rails application (it really doesn't make much sense to add comment
to `User` model for example).